requirements/local.txt
-r base.txt
django-debug-toolbar>=1.4
ipython>=4.1.1
